seed changes for postgres

+++ database/seeds/00_cleanup.js
@@ -1,5 +1,7 @@
 exports.seed = async function(knex) {
-  await knex("class").truncate();
-  await knex("user").truncate();
-  await knex("instructor").truncate();
+  if (process.env.NODE_ENV != "production") {
+    await knex("class").truncate();
+    await knex("user").truncate();
+    await knex("instructor").truncate();
+  }
 };
